// sdk/src/generated/types/DynamicRoyalty.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
import {
  DynamicRoyaltyPriceLinear,
  dynamicRoyaltyPriceLinearBeet,
} from './DynamicRoyaltyPriceLinear'
export type DynamicRoyalty = {
  version: number
  kind: number
  overrideRoyaltyBp: beet.COption<number>
  kindPriceLinear: beet.COption<DynamicRoyaltyPriceLinear>
  reserved0: number[] /* size: 32 */
  reserved1: number[] /* size: 32 */
  reserved2: number[] /* size: 32 */
  reserved3: number[] /* size: 32 */
}

/**
 * @category userTypes
 * @category generated
 */
export const dynamicRoyaltyBeet =
  new beet.FixableBeetArgsStruct<DynamicRoyalty>(
    [
      ['version', beet.u8],
      ['kind', beet.u8],
      ['overrideRoyaltyBp', beet.coption(beet.u16)],
      ['kindPriceLinear', beet.coption(dynamicRoyaltyPriceLinearBeet)],
      ['reserved0', beet.uniformFixedSizeArray(beet.u8, 32)],
      ['reserved1', beet.uniformFixedSizeArray(beet.u8, 32)],
      ['reserved2', beet.uniformFixedSizeArray(beet.u8, 32)],
      ['reserved3', beet.uniformFixedSizeArray(beet.u8, 32)],
    ],
    'DynamicRoyalty'
  )